users_service/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
pydantic==1.10.12
email-validator==2.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
